# docs/reference/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {"title":"AreaFlow API","version":"1.0.0","license":{"name":"Apache-2.0","identifier":"Apache-2.0"}},
  "servers": [{"url":"/api/v1"}],
  "security": [{"bearerAuth":[]},{"sessionCookie":[]}],
...
  "paths": {
...
    "/api/v1/auth/oidc/callback":{"get":{"security":[],"responses":{"302":{"description":"Create web session"},"401":{"$ref":"#/components/responses/Problem"}}}},
...
  },
  "components": {
    "securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"},"sessionCookie":{"type":"apiKey","in":"cookie","name":"areaflow_session"}},
...
    "responses":{"Problem":{"description":"RFC 7807 problem details","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/Problem"}}}}},
    "schemas":{
      "Problem":{"type":"object","required":["type","title","status","detail"],"properties":{"type":{"type":"string"},"title":{"type":"string"},"status":{"type":"integer"},"detail":{"type":"string"},"error":{"type":"string"},"request_id":{"type":"string"}}},
...
    }
  }
}